requirements.txt
transformers
requests
black
fastapi
uvicorn
pillow
jupyter
torch
python-multipart